+++ src/ts/events/hacks/HackTimer.ts
@@ -66,7 +66,11 @@ class HackTimer {
                 new Cryptogram(threatLevel);
                 break;
             case 1:
-                new HiddenPasswords(threatLevel);
+                if (Settings.isSettingEnabled("poor-eyesight-features")) {
+                    HackTimer.createHack();
+                } else {
+                    new HiddenPasswords(threatLevel);
+                }
                 break;
             case 2:
                 new NumberMultiples(threatLevel);